Refactor pkgs
