+++ pkg/app/server.go
@@ -1,4 +1,4 @@
-package discovery
+package app
 
 import "context"
